workers/requirements.txt
# MongoDB driver
pymongo==4.6.1

# Cron-like scheduling
APScheduler==3.10.4

# HTTP requests for health checks
requests==2.31.0

# HTML parsing for SEO analysis
beautifulsoup4==4.12.3


# Environment variables
python-dotenv==1.0.0

# SSL certificates (for validation)
certifi==2024.2.2
